完成剩下几个小系统重构，其实不需要改夺少东西

+++ Assets/Scripts/Core/Combat/Replication/WeaponSnapshot.cs
@@ -51,7 +51,7 @@ public struct WeaponSnapshot : INetworkSerializable
         where T : IReaderWriter
     {
         serializer.SerializeValue(ref EventCount);
-        int count=Math.Min(EventCount,MaxEvents);
+        int count=Math.Min((int)EventCount,MaxEvents);
         for(int i=0;i<count;i++)
         {
             ShotData shot=GetEvent(i);
